feat: extract content from page and provide content separately
