+++ .changeset/short-timers-act.md
@@ -1,0 +1,5 @@
+---
+'@dweber019/backstage-plugin-missing-entity': patch
+---
+
+Extract content from page and provide content separately.
