+++ plugins/missing-entity/src/components/MissingEntityPage/index.ts
@@ -1,1 +1,2 @@
 export { MissingEntityPage } from './MissingEntityPage';
+export { MissingEntityContent } from './MissingEntityContent';
