+++ plugins/missing-entity/src/index.ts
@@ -1,5 +1,6 @@
 export {
   missingEntityPlugin,
   MissingEntityPage,
+  MissingEntityContent,
 } from './plugin';
 export * from './components/MissingEntityWarning';
